+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.58.1] - 2023-02-15
+
+### Changed
+
+- Ability to add multiple actions in a `List::Item::Component`
+- Ability to add content in the middle of a `List::Item`
+
 ## [0.58.0] - 2023-02-11
 
 ### Changed